fixing func naming conflict
